Fix issue 460
