+++ setup.cfg
@@ -19,4 +19,4 @@ ignore =
 
 [codespell]
 quiet-level = 2
-skip = poetry.lock,./.git,./.mypy_cache,./dist,*.csv
+skip = poetry.lock,./.git,./.mypy_cache,./dist,*.csv,./env
